src/app/auth-guard/auth.guard.ts
import { AuthService } from '../auth-service/auth.service';

import { Injectable } from '@angular/core';
import {
  ActivatedRouteSnapshot,
  CanActivate,
  Router,
  RouterStateSnapshot
} from '@angular/router';

import 'rxjs/add/operator/do';
import 'rxjs/add/operator/map';
import { Observable } from 'rxjs/Observable';

@Injectable()
export class AuthGuard implements CanActivate {
  constructor(private auth: AuthService, private router: Router) {}

  canActivate(
    next: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ): Observable<boolean> | boolean {
    return this.auth.user$
      .map((user) => (user ? true : false))
      .do((authenticated) => {
        if (!authenticated) {
          this.router.navigate(['/login']);
        }
      });
  }
}
